Split production modules into different concerns and in different packages
