+++ app/categorytree/RefreshableCategoryTree.java
@@ -1,4 +1,4 @@
-package controllers;
+package categorytree;
 
 import io.sphere.sdk.categories.Category;
 import io.sphere.sdk.categories.CategoryTree;
