+++ app/ctpclient/CtpClientProductionModule.java
@@ -1,4 +1,4 @@
-package inject;
+package ctpclient;
 
 import com.google.inject.AbstractModule;
 import io.sphere.sdk.client.PlayJavaSphereClient;
@@ -7,8 +7,7 @@ import io.sphere.sdk.client.SphereClient;
 import javax.inject.Singleton;
 
 /**
- * Configuration for the Guice {@link com.google.inject.Injector} which
- * shall be used in production and integration tests.
+ * Configuration for the Guice {@link com.google.inject.Injector} which shall be used in production.
  */
 public class CtpClientProductionModule extends AbstractModule {
