+++ app/inject/ApplicationProductionModule.java
@@ -2,16 +2,13 @@ package inject;
 
 import com.google.inject.AbstractModule;
 import common.cms.CmsService;
-import common.controllers.ReverseRouter;
 import common.i18n.I18nResolver;
 import common.templates.TemplateService;
-import pages.ReverseRouterImpl;
 
 import javax.inject.Singleton;
 
 /**
- * Configuration for the Guice {@link com.google.inject.Injector} which
- * shall be used in production and integration tests.
+ * Configuration for the Guice {@link com.google.inject.Injector} which shall be used in production.
  */
 public class ApplicationProductionModule extends AbstractModule {
 
@@ -20,6 +17,5 @@ public class ApplicationProductionModule extends AbstractModule {
         bind(I18nResolver.class).toProvider(I18nResolverProvider.class).in(Singleton.class);
         bind(TemplateService.class).toProvider(TemplateServiceProvider.class).in(Singleton.class);
         bind(CmsService.class).toProvider(CmsServiceProvider.class).in(Singleton.class);
-        bind(ReverseRouter.class).toInstance(new ReverseRouterImpl());
     }
 }
